added easy_mpl into dependency
