+++ ai4water/eda/_main.py
@@ -7,16 +7,16 @@ import pandas as pd
 import scipy.stats as stats
 import matplotlib.pyplot as plt
 import matplotlib.ticker as ticker
+from easy_mpl import bar_chart
 
 try:
     import seaborn as sns
 except ModuleNotFoundError:
     sns = None
 
-from .utils import pac_yw, auto_corr, plot_autocorr
 from .utils import _missing_vals
+from .utils import pac_yw, auto_corr, plot_autocorr
 from ai4water.utils.visualizations import Plot
-from ai4water.utils.easy_mpl import bar_chart
 from ai4water.utils.utils import find_tot_plots
 from ai4water.preprocessing import Transformation
 from ai4water.utils.utils import dict_to_file, dateandtime_now, ts_features
