+++ ai4water/postprocessing/explain/_shap.py
@@ -9,6 +9,7 @@ except ModuleNotFoundError:
 import scipy as sp
 import numpy as np
 import pandas as pd
+from easy_mpl import imshow
 import matplotlib.pyplot as plt
 
 try:
@@ -16,7 +17,6 @@ try:
 except ModuleNotFoundError:
     K = None
 
-from ai4water.utils.easy_mpl import imshow
 from ai4water.backend import sklearn_models
 from ._explain import ExplainerMixin
 from .utils import convert_ai4water_model
