+++ ai4water/postprocessing/visualize/_main.py
@@ -5,15 +5,19 @@ import warnings
 from typing import Union
 
 import numpy as np
+from easy_mpl import imshow
 import matplotlib.pyplot as plt
 from sklearn.tree import plot_tree
 
 from ai4water.backend import tf, keras
 
-import ai4water.keract_mod as keract
+if tf is not None:
+    import ai4water.keract_mod as keract
+else:
+    keract = None
+
 from ai4water.utils.plotting_tools import Plots
 from ai4water.utils.utils import maybe_three_outputs
-from ai4water.utils.easy_mpl import imshow
 
 from ..utils import choose_examples
