+++ ai4water/pytorch_models.py
@@ -5,19 +5,18 @@ from typing import Any
 
 import numpy as np
 import matplotlib as mpl
+from easy_mpl import imshow
 import matplotlib.pyplot as plt
 
-from ai4water import Model
-from ai4water.utils.utils import dateandtime_now, plot_activations_along_inputs
-from ai4water.backend import torch
+from .main import Model
+from .utils.utils import dateandtime_now, plot_activations_along_inputs
+from .backend import torch
 
 if torch is not None:
-    from ai4water.models.torch import IMVTensorLSTM
-    from ai4water.models.torch import HARHN
+    from .models.torch import IMVTensorLSTM
+    from .models.torch import HARHN
 else:
     HARHN, IMVTensorLSTM = None, None
-
-from ai4water.utils.easy_mpl import imshow
 
 
 class HARHNModel(Model):
@@ -133,7 +132,6 @@ class IMVModel(HARHNModel):
                title="Importance of features and timesteps",
                annotate=annotate,
                show=False)
-
 
         plt.savefig(os.path.join(path, f'acts_{name}'), dpi=400, bbox_inches='tight')
