+++ ai4water/utils/__init__.py
@@ -1,5 +1,4 @@
 
+from .taylor_diagram import taylor_plot
 from .utils import make_model, prepare_data
 from .visualizations import ProcessResults, murphy_diagram, fdc_plot
-from .taylor_diagram import taylor_plot
-from .easy_mpl import plot, regplot, imshow, bar_chart
