+++ ai4water/utils/plotting_tools.py
@@ -3,8 +3,8 @@ from typing import Union
 
 import numpy as np
 import pandas as pd
+from easy_mpl import imshow
 import matplotlib.pyplot as plt
-
 
 # TODO
 # rank histogram, reliability diagram, ROC curve
@@ -73,8 +73,6 @@ class Plots(object):
                 show=False,
                 **kwargs):
 
-        from .easy_mpl import imshow
-
         assert np.ndim(img) == 2, "can not plot {} with shape {} and ndim {}".format(label, img.shape, np.ndim(img))
 
         axis, im = imshow(img,
@@ -132,7 +130,7 @@ class Plots(object):
         return save_or_show(self.path, *args, **kwargs)
 
     def plot2d_act_for_a_sample(self, activations, sample=0, save: bool = False, name: str = None):
-        from ai4water.utils.easy_mpl import init_subplots
+        from ai4water.utils.visualizations import init_subplots
 
         fig, axis = init_subplots(height=8)
         # for idx, ax in enumerate(axis):
