+++ requirements.txt
@@ -7,3 +7,5 @@ plotly
 joblib
 requests
 matplotlib
+
+easy_mpl
